Buddy_alloc: Fix printf warnings for 32bit builds
The type of an enum is not considered
for printf-fmt checking, at least for longs,
i.e. there are printf warnings.
Thus add explicit casts to "unsigned long".
This seems like an issue in gcc <= 7, i.e.
no warning with gcc8, as with clang.

+++ src/kern/buddy_alloc.cpp
@@ -222,13 +222,13 @@ void
 Buddy_t_base<A,B,M>::dump() const
 {
   unsigned long total = 0;
-  printf("Buddy_alloc [%ld,%ld]\n", Min_size, Num_sizes);
+  printf("Buddy_alloc [%ld,%ld]\n", (unsigned long)Min_size, (unsigned long)Num_sizes);
   for (unsigned i = 0; i < Num_sizes; ++i)
     {
       unsigned long c = 0;
       unsigned long avail = 0;
       B_list::Const_iterator h = _free[i].begin();
-      printf("  [%ld] %p(%lu)", Min_size << i, *h, h != _free[i].end() ? h->index : 0UL);
+      printf("  [%ld] %p(%lu)", (unsigned long)Min_size << i, *h, h != _free[i].end() ? h->index : 0UL);
       while (h != _free[i].end())
 	{
 	  ++h;
